+++ stories/02-Design System/01-General/Button/Button.stories.tsx
@@ -5,9 +5,9 @@ import {argTypes} from './data';
 import {Template} from './Template';
 import {PrimaryButtonTest} from './test-components/PrimaryButtonTest.tsx';
 import {DefaultButtonTest} from './test-components/DefaultButtonTest.tsx';
+import {TertiaryButtonTest} from './test-components/TertiaryButtonTest.tsx';
 import {DangerButtonTest} from './test-components/DangerButtonTest.tsx';
 import {TextButtonTest} from './test-components/TextButtonTest.tsx';
-import {LinkButtonTest} from './test-components/LinkButtonTest.tsx';
 import {SegmentedButtonTest} from './test-components/SegmentedButtonTest.tsx';
 import {IconButtonTest} from './test-components/IconButtonTest.tsx';
 import {IconSizeButtonTest} from './test-components/IconSizeButtonTest.tsx';
@@ -34,9 +34,9 @@ export const ChromaticTest: Story = {
         <>
             <PrimaryButtonTest />
             <DefaultButtonTest />
+            <TertiaryButtonTest />
             <DangerButtonTest />
             <TextButtonTest />
-            <LinkButtonTest />
             <SegmentedButtonTest />
             <IconButtonTest />
             <IconSizeButtonTest />