+++ stories/02-Design System/03-Navigation/Menu/examples/Menu/menuHeaderButton.tsx
@@ -19,12 +19,7 @@ const App = () => {
                     title="View options"
                     onCloseClick={() => console.log('on click close')}
                     segmentedButton={
-                        <KitButton
-                            type="segmented"
-                            segmentedColor="green"
-                            segmentedActived
-                            icon={<FontAwesomeIcon icon={faHouse} />}
-                        />
+                        <KitButton type="segmented" color="green" active icon={<FontAwesomeIcon icon={faHouse} />} />
                     }
                     primaryInput={<KitInput placeholder="Opération #1" />}
                 />
@@ -32,12 +27,7 @@ const App = () => {
                     title="View options"
                     onCloseClick={() => console.log('on click close')}
                     segmentedButton={
-                        <KitButton
-                            type="segmented"
-                            segmentedColor="green"
-                            segmentedActived
-                            icon={<FontAwesomeIcon icon={faHouse} />}
-                        />
+                        <KitButton type="segmented" color="green" active icon={<FontAwesomeIcon icon={faHouse} />} />
                     }
                     primaryInput={<KitInput placeholder="Opération #1" />}
                     secondaryInput={<KitInput placeholder="Description" />}